wip: skip admin import specs
