+++ apps/cli/src/use-cases/export/export-seed-data.ts
@@ -7,8 +7,18 @@ export const exportSeedData = async () => {
   const atomsData = await exportAtoms()
   const tagsData = await exportTags()
   const seedTypesData = await exportSeedTypes()
+
   // We'll want to sort the data so diff is minimized
-  const sortedAtomsData = atomsData.sort((a, b) => a.name.localeCompare(b.name))
+  const sortedAtomsData = atomsData
+    // Sort by atom name
+    .sort((a, b) => a.name.localeCompare(b.name))
+    // Sort the allowed children data as well
+    .map((atom) => ({
+      ...atom,
+      allowedChildren: atom.allowedChildren.sort((a, b) =>
+        a.name.localeCompare(b.name),
+      ),
+    }))
 
   const seedData: Omit<
     ExportedData,
