+++ libs/shared/abstract/core/src/data/index.ts
@@ -3,3 +3,4 @@ export * from './atom.interface'
 export * from './data.interface'
 export * from './store.interface'
 export * from './tag.interface'
+export * from './type.interface'
