+++ libs/shared/abstract/core/src/domain/type/types/react-node-type/react-node-type.dto.interface.ts
@@ -1,11 +1,3 @@
-import { OGM_TYPES } from '@codelab/shared/abstract/codegen'
 import { ReactNodeTypeFragment } from '../../fragments'
-import { IBaseTypeExportFields } from '../base-type'
 
 export type IReactNodeTypeDTO = ReactNodeTypeFragment
-
-export type IReactNodeTypeExport = Pick<
-  OGM_TYPES.ReactNodeType,
-  IBaseTypeExportFields
-> &
-  Required<Pick<OGM_TYPES.ReactNodeType, '__typename'>>
