+++ libs/shared/abstract/core/src/domain/type/types/render-props-type/render-props-type.dto.interface.ts
@@ -1,11 +1,3 @@
-import { OGM_TYPES } from '@codelab/shared/abstract/codegen'
 import { RenderPropsTypeFragment } from '../../fragments'
-import { IBaseTypeExportFields } from '../base-type'
 
 export type IRenderPropsTypeDTO = RenderPropsTypeFragment
-
-export type IRenderPropsTypeExport = Pick<
-  OGM_TYPES.RenderPropsType,
-  IBaseTypeExportFields
-> &
-  Required<Pick<OGM_TYPES.RenderPropsType, '__typename'>>
